group commands to files
random_apis.py -> fun.py
move urban and weather commands to info.py
